crunch: other way to detect if a path is a file or a directory
readdir() doesn't support file detection on all file systems,
so we use stat() as a fallback.

This makes the crunch tool works with files stored over
the network on NFS shares on MacOS.

We also report if the current code fails to detect if a
path is a file or a directory, just in case some other
corner case surfaces.